customuser/requirements.txt
appdirs==1.4.3
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.0
django-allauth==0.41.0
filelock==3.0.12
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.4.0
oauthlib==3.1.0
pipenv==2018.11.26
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
virtualenv==20.0.17
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
zipp==3.1.0
